+++ framework/services-comun-status/modules/services/status-backend/auth/interface.ts
@@ -1,4 +1,14 @@
+/**
+ * Editor: Bixus
+ * Fecha: Mon, 17 Aug 2026 06:32:04 GMT
+ * Hash: 0f3dda68776882c821f62d874a15f76d
+ * Versión: 2026.8.17+1-bixus
+ * Anterior: 2026.8.13+2-bixus
+ * Proyecto: https://github.com/meteored-status/svc-status.git
+ */
+
 export type EPermission =
+    "status.admin"                      |
     "status.panel.access"               |
     "status.panel.guardian.view"        |
     "status.resources.list"             |
@@ -13,6 +23,13 @@ export type EPermission =
     "status.subscription.toggle"        |
     "status.component.delete"           |
     "status.log.list"                   |
-    "status.elastic.list"               |
-    "status.mysql.list"
+    "status.user.list"                  |
+    "status.user.edit"                  |
+    "status.user.delete"                |
+    "status.rol.list"                   |
+    "status.rol.edit"                   |
+    "status.rol.delete"                 |
+    "status.dpto.list"                  |
+    "status.dpto.edit"                  |
+    "status.dpto.delete"
 ;